Refactor binary
