+++ src/binary/help/lib.txt
@@ -1,0 +1,17 @@
+Hello, welcome to Dune's help macro!
+
+Dune offers the following builtin libraries:
+1. `math`: a library with several math helper functions.
+2. `time`: a library with time related functions.
+3. `rand`: a library with random number generation functions.
+4. `os`: a library with operating system related functions.
+5. `fs`: a library with file system related functions.
+6. `fn`: a library with functional programming constructs.
+7. `fmt`: a library with color, formatting, and other text functions.
+8. `widget`: a library for creating widgets on the console.
+9. `console`: a library for manipulating the terminal console.
+
+To see all the different functions and constants for each library,
+simply print the library itself!
+
+$ echo math
