+++ src/binary/mod.rs
@@ -1,0 +1,3 @@
+mod init_environment;
+
+pub use init_environment::init_environment;
